+++ packages/core/src/components/tree-view/tree-view.tsx
@@ -129,7 +129,7 @@ export class TreeView implements ComponentInterface {
     event.preventDefault();
   };
 
-  // eslint-disable-next-line @stencil-community/own-props-must-be-private -- Omdat this als TreeView instance aan Functionele Components wordt gegeven
+  // eslint-disable-next-line stencil/own-props-must-be-private -- accessed by tree-item.tsx via owner reference
   itemClick = (event: MouseEvent, ancestors: TreeViewItem[], item: TreeViewItem) => {
     if (!(event.target instanceof HTMLElement)) {
       return;